ci: split calendar (r) e2e shard into 10 parallel shards (n–w) (#936)
* ci: split calendar (r) e2e shard into 10 shards (n,o,p,q,r,s,t,u,v,w)

The r/ calendar regression suite (21 specs) ran serially as a single
matrix shard, ~56 min of tests + ~6 min setup (~62 min wall-clock) — the
long pole of the PR matrix. Split it across the previously-empty n/o/p/q
slots plus s/t/u/v/w so each shard runs ~5-7 min of tests in parallel.

- Distribute the 21 spec files across n,o,p,q,r,s,t,u,v,w (~2 files each;
  r keeps the heaviest, calendar-ui-enhancements).
- Move the shared calendar-ui-enhancements.page.ts up to the
  backend-configuration-pn/ dir (it has no relative imports); repoint all
  21 specs to ../calendar-ui-enhancements.page (no duplication).
- Co-locate both attachments specs in shard w with their fixtures/
  (referenced via __dirname).
- Add n,o,p,q,s,t,u,v,w to the matrix in both dotnet-core-pr.yml and
  dotnet-core-master.yml. New shards are non-'a', so they follow the same
  no-SQL-seed path r already used.

Balance is by test-count proxy (CI emits only a run summary, not per-file
times); will fine-tune from the new per-shard summaries — and split
calendar-ui-enhancements if its shard exceeds ~7 min.


* ci: rebalance shard w (attachments) — move copy→p, weekdays-repeat→o

First run measured shard w at 7.9m test time (the file-upload attachments
specs are heavier per-test than their count suggested), just over the
5–7 min target. The attachments specs must share fixtures/, so move the
two non-attachments files off w onto the lightest shards: copy→p,
weekdays-repeat→o. Expected w ≈ 6m; p/o ≈ 5.9m. r (ui-enhancements)
came in at 5.6m, so no file-split needed there.


* ci: share calendar-attachments fixtures across shards (fix copy ENOENT)

P04 in calendar-copy (now on shard p) also uploads an attachment fixture
via path.resolve(__dirname, 'fixtures/...'), but fixtures/ lived only in
shard w → ENOENT. Move fixtures/ up to the shared backend-configuration-pn/
dir and point all three fixture-using specs (calendar-copy,
calendar-attachments, calendar-attachments-quota) at ../fixtures/calendar-
attachments. No duplication; copy stays on p, attachments on w.


---------
